requirements.txt
asgiref==3.11.0
Django==6.0
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
drf-yasg==1.21.14
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.22
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
